[IMP] dropbox: show a message if we might overwrite remote wallet
Purpose
=======
If you login from 2 different devices, if you made changes changes
in the wallet on those 2 devices and then upload them to Dropbox,
one wallet might overwrite the other.

For that purpose, we store the last file hash we have from Dropbox,
and if during the upload this hash is different, it means that our
local wallet was not up-to-date and that we might loose data.